fastapi_whisper_app/requirements.txt

fastapi
uvicorn[standard]
python-multipart
openai-whisper
transformers
torchaudio
pydub
moviepy
langdetect
nltk
spacy
torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
